:smile_cat:
adding main core functionality, there are plans for a lot of other areas of recon that i personally have seen are important, the drawback is PHP has no pure PHP way of handling SMB connections, and i refuse to install anything with composer unless i absolutely have to, this reduces bloat, and ensure that it will run out of the box with a portable version of PHP, once i figure out how to interact at the socket level with an SMB share, that will be added. SQL stuff will also be added.