😸
adding in functions and skeletons to search registry hives + the custom implementation of the SMB/NBT protocols in PHP [:(]

Added more language on what this tool does, its detection (if any) when used on a system from a SOC perspective, COM events appear to be the only tell of this tool being deployed for now. Network connections and such will likely raise a TON of alarms, if it doesn't, that's not good.